refactor(infra): split docker_runner into a docker package
Decompose the 1291-line infra/docker_runner.py god-module into an
infra/docker/ package along concern seams, keeping DockerRunner as the
facade at infra.docker_runner with its constructor and public method
signatures unchanged:

- command.py: pure docker-run command builders (env forwarding, mounts,
  GPU/shm flags, image ref, package-dispatch bootstrap).
- lifecycle.py: container process execution. The block-until-exit path is
  a separable launch() (detach-capable) plus wait_to_completion(), with
  the stdout-silence + wall-clock watchdog isolated inside the wait path
  only, leaving a clean seam for a future server mode (no server dispatch
  is built here).
- exchange.py: exchange-dir lifecycle, result read, artefact rescue.
- diagnostics.py: container failure classification (log tail, error
  payload, error mapping).

Extract the ~90-line dependency-import probe previously embedded as a bash
heredoc in container_entrypoint.sh into a real package-data module
(infra/_container/probe_imports.py) that lint/mypy cover and unit tests
import directly. The shell script invokes it from the mounted package
source and fails loudly with a clear message if it is absent.

Make the host HuggingFace cache mount configurable via LLEM_DOCKER_HF_CACHE
(default $HOME/.cache/huggingface), mirroring LLEM_DOCKER_SHM_SIZE.

Behavior is frozen: same env/mounts/flags, watchdog thresholds and
messages, error classification, rescue semantics, and dep-isolation. Docker
tests keep their coverage; syscall monkeypatch targets move to the concern
modules at the use site.
